+++ include/UDRefl/Basic.h
@@ -59,9 +59,6 @@ namespace Ubpa::UDRefl {
 	class ObjectView;
 	class SharedObject;
 
-	// pointer const array type (pointer is const, and pointer to non - const / referenced object)
-	using ArgPtrBuffer = void* const*;
-
 	template<typename T>
 	struct IsObjectOrView {
 	private: