+++ include/UDRefl/ReflMngr.h
@@ -271,8 +271,7 @@ namespace Ubpa::UDRefl {
 			ObjectView obj,
 			Name method_name,
 			void* result_buffer = nullptr,
-			std::span<const Type> argTypes = {},
-			ArgPtrBuffer argptr_buffer = nullptr,
+			ArgsView args = {},
 			MethodFlag flag = MethodFlag::All,
 			std::pmr::memory_resource* temp_args_rsrc = Mngr->GetTemporaryResource()) const;
 
@@ -281,17 +280,15 @@ namespace Ubpa::UDRefl {
 			Name method_name,
 			std::pmr::memory_resource* rst_rsrc,
 			std::pmr::memory_resource* temp_args_rsrc,
-			std::span<const Type> argTypes = {},
-			ArgPtrBuffer argptr_buffer = nullptr,
+			ArgsView args = {},
 			MethodFlag flag = MethodFlag::All) const;
 
 		SharedObject Invoke(
 			ObjectView obj,
 			Name method_name,
-			std::span<const Type> argTypes = {},
-			ArgPtrBuffer argptr_buffer = nullptr,
+			ArgsView args = {},
 			MethodFlag flag = MethodFlag::All) const
-		{ return MInvoke(obj, method_name, &object_resource, &temporary_resource, argTypes, argptr_buffer, flag); }
+		{ return MInvoke(obj, method_name, &object_resource, &temporary_resource, args, flag); }
 
 		// -- template --
 
@@ -302,8 +299,7 @@ namespace Ubpa::UDRefl {
 		T BInvokeRet(
 			ObjectView obj,
 			Name method_name,
-			std::span<const Type> argTypes = {},
-			ArgPtrBuffer argptr_buffer = nullptr,
+			ArgsView args = {},
 			MethodFlag flag = MethodFlag::All,
 			std::pmr::memory_resource* temp_args_rsrc = Mngr->GetTemporaryResource()) const;
 
@@ -338,16 +334,16 @@ namespace Ubpa::UDRefl {
 		bool IsMoveConstructible(Type type) const;
 		bool IsDestructible     (Type type) const;
 
-		bool Construct(ObjectView obj, std::span<const Type> argTypes = {}, ArgPtrBuffer argptr_buffer = nullptr) const;
+		bool Construct(ObjectView obj, ArgsView args = {}) const;
 		void Destruct (ObjectView obj) const;
 
-		ObjectView   MNew       (Type      type, std::pmr::memory_resource* rsrc, std::span<const Type> argTypes = {}, ArgPtrBuffer argptr_buffer = nullptr) const;
-		SharedObject MMakeShared(Type      type, std::pmr::memory_resource* rsrc, std::span<const Type> argTypes = {}, ArgPtrBuffer argptr_buffer = nullptr) const;
-		bool         MDelete    (ObjectView obj, std::pmr::memory_resource* rsrc                                                                           ) const;
+		ObjectView   MNew       (Type      type, std::pmr::memory_resource* rsrc, ArgsView args = {}) const;
+		SharedObject MMakeShared(Type      type, std::pmr::memory_resource* rsrc, ArgsView args = {}) const;
+		bool         MDelete    (ObjectView obj, std::pmr::memory_resource* rsrc                    ) const;
 
-		ObjectView   New       (Type      type, std::span<const Type> argTypes = {}, ArgPtrBuffer argptr_buffer = nullptr) const;
-		SharedObject MakeShared(Type      type, std::span<const Type> argTypes = {}, ArgPtrBuffer argptr_buffer = nullptr) const;
-		bool         Delete    (ObjectView obj                                                                           ) const;
+		ObjectView   New       (Type      type, ArgsView args = {}) const;
+		SharedObject MakeShared(Type      type, ArgsView args = {}) const;
+		bool         Delete    (ObjectView obj                    ) const;
 
 		// -- template --