+++ src/core/InvokeUtil.h
@@ -48,7 +48,7 @@ namespace Ubpa::UDRefl::details {
 	bool IsRefCompatible(std::span<const Type> paramTypes, std::span<const TypeID> argTypeIDs);
 
 	bool IsRefConstructible(Type paramType, std::span<const Type> argTypes);
-	bool RefConstruct(ObjectView obj, std::span<const Type> argTypes, ArgPtrBuffer argptr_buffer);
+	bool RefConstruct(ObjectView obj, ArgsView args);
 
 	class BufferGuard {
 	public:
@@ -95,8 +95,7 @@ namespace Ubpa::UDRefl::details {
 			bool is_priority,
 			std::pmr::memory_resource* rsrc,
 			std::span<const Type> paramTypes,
-			std::span<const Type> argTypes,
-			ArgPtrBuffer orig_argptr_buffer);
+			ArgsView args);
 
 		~NewArgsGuard();
 
@@ -107,14 +106,14 @@ namespace Ubpa::UDRefl::details {
 
 		ArgsView GetArgsView() const noexcept {
 			assert(IsCompatible());
-			return args;
+			return new_args;
 		}
 
 	private:
 		bool is_compatible{ false };
 		BufferGuard buffer;
 		std::span<ArgInfo> nonptr_arg_infos;
-		ArgsView args;
+		ArgsView new_args;
 		BufferGuard type_buffer;
 	};
 }