+++ src/core/ReflMngr.cpp
@@ -159,8 +159,7 @@ namespace Ubpa::UDRefl::details {
 		ObjectView obj,
 		Name method_name,
 		void* result_buffer,
-		std::span<const Type> argTypes,
-		ArgPtrBuffer argptr_buffer,
+		ArgsView args,
 		MethodFlag flag)
 	{
 		assert(obj.GetType().GetCVRefMode() == CVRefMode::None);
@@ -179,7 +178,7 @@ namespace Ubpa::UDRefl::details {
 				if (enum_contain(MethodFlag::Priority, iter->second.methodptr.GetMethodFlag())) {
 					NewArgsGuard guard{
 						is_priority, args_rsrc,
-						iter->second.methodptr.GetParamList(), argTypes, argptr_buffer
+						iter->second.methodptr.GetParamList(), args
 					};
 					if (!guard.IsCompatible())
 						continue;
@@ -193,7 +192,7 @@ namespace Ubpa::UDRefl::details {
 				if (iter->second.methodptr.GetMethodFlag() == MethodFlag::Const) {
 					NewArgsGuard guard{
 						is_priority, args_rsrc,
-						iter->second.methodptr.GetParamList(), argTypes, argptr_buffer
+						iter->second.methodptr.GetParamList(), args
 					};
 					if (!guard.IsCompatible())
 						continue;
@@ -208,7 +207,7 @@ namespace Ubpa::UDRefl::details {
 			auto rst = BInvoke(
 				is_priority, args_rsrc,
 				ObjectView{ base, baseinfo.StaticCast_DerivedToBase(obj.GetPtr()) },
-				method_name, result_buffer, argTypes, argptr_buffer,
+				method_name, result_buffer, args,
 				flag
 			);
 			if (rst)
@@ -224,8 +223,7 @@ namespace Ubpa::UDRefl::details {
 		ObjectView obj,
 		Name method_name,
 		std::pmr::memory_resource* rst_rsrc,
-		std::span<const Type> argTypes,
-		ArgPtrBuffer argptr_buffer,
+		ArgsView args,
 		MethodFlag flag)
 	{
 		assert(args_rsrc);
@@ -245,7 +243,7 @@ namespace Ubpa::UDRefl::details {
 				if (enum_contain(MethodFlag::Priority, iter->second.methodptr.GetMethodFlag())) {
 					NewArgsGuard guard{
 						is_priority, args_rsrc,
-						iter->second.methodptr.GetParamList(), argTypes, argptr_buffer
+						iter->second.methodptr.GetParamList(), args
 					};
 
 					if (!guard.IsCompatible())
@@ -293,7 +291,7 @@ namespace Ubpa::UDRefl::details {
 				if (iter->second.methodptr.GetMethodFlag() == MethodFlag::Const) {
 					NewArgsGuard guard{
 						is_priority, args_rsrc,
-						iter->second.methodptr.GetParamList(), argTypes, argptr_buffer
+						iter->second.methodptr.GetParamList(), args
 					};
 
 					if (!guard.IsCompatible())
@@ -340,7 +338,7 @@ namespace Ubpa::UDRefl::details {
 			auto rst = MInvoke(
 				is_priority, args_rsrc,
 				ObjectView{ base, baseinfo.StaticCast_DerivedToBase(obj.GetPtr()) },
-				method_name, rst_rsrc, argTypes, argptr_buffer,
+				method_name, rst_rsrc, args,
 				flag
 			);
 			if (rst.GetType())
@@ -843,8 +841,8 @@ bool ReflMngr::AddMethodAttr(Type type, Name name, Attr attr) {
 	return true;
 }
 
-SharedObject ReflMngr::MMakeShared(Type type, std::pmr::memory_resource* rsrc, std::span<const Type> argTypes, ArgPtrBuffer argptr_buffer) const {
-	ObjectView obj = MNew(type, rsrc, argTypes, argptr_buffer);
+SharedObject ReflMngr::MMakeShared(Type type, std::pmr::memory_resource* rsrc, ArgsView args) const {
+	ObjectView obj = MNew(type, rsrc, args);
 
 	if (!obj.GetType().Valid())
 		return {};
@@ -1077,8 +1075,7 @@ Type ReflMngr::BInvoke(
 	ObjectView obj,
 	Name method_name,
 	void* result_buffer,
-	std::span<const Type> argTypes,
-	ArgPtrBuffer argptr_buffer,
+	ArgsView args,
 	MethodFlag flag,
 	std::pmr::memory_resource* temp_args_rsrc) const
 {
@@ -1108,10 +1105,10 @@ Type ReflMngr::BInvoke(
 	if (!obj.GetPtr())
 		flag = enum_within(flag, MethodFlag::Static);
 
-	if (auto priority_rst = details::BInvoke(true, temp_args_rsrc, rawObj, method_name, result_buffer, argTypes, argptr_buffer, flag))
+	if (auto priority_rst = details::BInvoke(true, temp_args_rsrc, rawObj, method_name, result_buffer, args, flag))
 		return priority_rst;
 
-	return details::BInvoke(false, temp_args_rsrc, rawObj, method_name, result_buffer, argTypes, argptr_buffer, flag);
+	return details::BInvoke(false, temp_args_rsrc, rawObj, method_name, result_buffer, args, flag);
 }
 
 SharedObject ReflMngr::MInvoke(
@@ -1119,8 +1116,7 @@ SharedObject ReflMngr::MInvoke(
 	Name method_name,
 	std::pmr::memory_resource* rst_rsrc,
 	std::pmr::memory_resource* temp_args_rsrc,
-	std::span<const Type> argTypes,
-	ArgPtrBuffer argptr_buffer,
+	ArgsView args,
 	MethodFlag flag) const
 {
 	assert(rst_rsrc);
@@ -1152,16 +1148,16 @@ SharedObject ReflMngr::MInvoke(
 	if (!obj.GetPtr())
 		flag = enum_within(flag, MethodFlag::Static);
 
-	if (auto priority_rst = details::MInvoke(true, temp_args_rsrc, rawObj, method_name, rst_rsrc, argTypes, argptr_buffer, flag); priority_rst.GetType().Valid())
+	if (auto priority_rst = details::MInvoke(true, temp_args_rsrc, rawObj, method_name, rst_rsrc, args, flag); priority_rst.GetType().Valid())
 		return priority_rst;
 
-	return details::MInvoke(false, temp_args_rsrc, rawObj, method_name, rst_rsrc, argTypes, argptr_buffer, flag);
+	return details::MInvoke(false, temp_args_rsrc, rawObj, method_name, rst_rsrc, args, flag);
 }
 
-ObjectView ReflMngr::MNew(Type type, std::pmr::memory_resource* rsrc, std::span<const Type> argTypes, ArgPtrBuffer argptr_buffer) const {
+ObjectView ReflMngr::MNew(Type type, std::pmr::memory_resource* rsrc, ArgsView args) const {
 	assert(rsrc);
 
-	if (!IsConstructible(type, argTypes))
+	if (!IsConstructible(type, args.Types()))
 		return {};
 
 	const auto& typeinfo = typeinfos.at(type);
@@ -1172,7 +1168,7 @@ ObjectView ReflMngr::MNew(Type type, std::pmr::memory_resource* rsrc, std::span<
 		return {};
 
 	ObjectView obj{ type, buffer };
-	bool success = Construct(obj, argTypes, argptr_buffer);
+	bool success = Construct(obj, args);
 	assert(success);
 
 	return obj;
@@ -1190,16 +1186,16 @@ bool ReflMngr::MDelete(ObjectView obj, std::pmr::memory_resource* rsrc) const {
 	return true;
 }
 
-ObjectView ReflMngr::New(Type type, std::span<const Type> argTypes, ArgPtrBuffer argptr_buffer) const {
-	return MNew(type, &object_resource, argTypes, argptr_buffer);
+ObjectView ReflMngr::New(Type type, ArgsView args) const {
+	return MNew(type, &object_resource, args);
 }
 
 bool ReflMngr::Delete(ObjectView obj) const {
 	return MDelete(obj, &object_resource);
 }
 
-SharedObject ReflMngr::MakeShared(Type type, std::span<const Type> argTypes, ArgPtrBuffer argptr_buffer) const {
-	return MMakeShared(type, &object_resource, argTypes, argptr_buffer);
+SharedObject ReflMngr::MakeShared(Type type, ArgsView args) const {
+	return MMakeShared(type, &object_resource, args);
 }
 
 bool ReflMngr::IsConstructible(Type type, std::span<const Type> argTypes) const {
@@ -1246,19 +1242,19 @@ bool ReflMngr::IsDestructible(Type type) const {
 	return false;
 }
 
-bool ReflMngr::Construct(ObjectView obj, std::span<const Type> argTypes, ArgPtrBuffer argptr_buffer) const {
+bool ReflMngr::Construct(ObjectView obj, ArgsView args) const {
 	auto target = typeinfos.find(obj.GetType());
 	if (target == typeinfos.end())
 		return false;
 	const auto& typeinfo = target->second;
 	auto [begin_iter, end_iter] = typeinfo.methodinfos.equal_range(NameIDRegistry::Meta::ctor);
-	if (begin_iter == end_iter && argTypes.empty())
+	if (begin_iter == end_iter && args.Types().empty())
 		return true;// trivial ctor
 	for (auto iter = begin_iter; iter != end_iter; ++iter) {
 		if (iter->second.methodptr.GetMethodFlag() == MethodFlag::Variable) {
 			details::NewArgsGuard guard{
 				false, &temporary_resource,
-				iter->second.methodptr.GetParamList(), argTypes, argptr_buffer
+				iter->second.methodptr.GetParamList(), args
 			};
 			if (!guard.IsCompatible())
 				continue;