+++ src/ext/Bootstrap/Bootstrap_helper_object.cpp
@@ -4,6 +4,13 @@ using namespace Ubpa;
 using namespace Ubpa::UDRefl;
 
 void Ubpa::UDRefl::ext::details::Bootstrap_helper_object() {
+	Mngr->RegisterType<ArgPtrBuffer>();
+
+	Mngr->RegisterType<ArgsView>();
+	Mngr->AddConstructor<ArgsView, ArgPtrBuffer, std::span<const Type>>();
+	Mngr->AddMethod<&ArgsView::Buffer>("Buffer");
+	Mngr->AddMethod<&ArgsView::Types>("Types");
+
 	Mngr->RegisterType<ObjectView>();
 	Mngr->AddConstructor<ObjectView, Type, void*>();
 	Mngr->AddConstructor<ObjectView, Type>();