+++ src/ext/Bootstrap/Bootstrap_helper_ptr.cpp
@@ -21,12 +21,6 @@ void Ubpa::UDRefl::ext::details::Bootstrap_helper_ptr() {
 
 	Mngr->RegisterType<ParamList>();
 
-	Mngr->RegisterType<ArgPtrBuffer>();
-
-	Mngr->RegisterType<ArgsView>();
-	Mngr->AddConstructor<ArgsView, ArgPtrBuffer, std::span<const Type>>();
-	Mngr->AddMethod<&ArgsView::Size>("Size");
-
 	Mngr->RegisterType<MethodPtr::Func>();
 	Mngr->RegisterType<MethodPtr>();
 	Mngr->AddConstructor<MethodPtr, MethodPtr::Func, MethodFlag, Type, ParamList>();